Update Power of four(Leetcode 342).java

+++ Bit Manipulation/Power of four(Leetcode 342).java
@@ -32,7 +32,7 @@ class Solution {
         return ((n&(n-1))==0) && isPerfectSquare(n);
     }
     public boolean isPerfectSquare(int n){
-        double root=Math.sqrt(n);
+        int root=(int)Math.sqrt(n);
         return root*root==n;
     }
 }
